Modularize AI service and update tests
